requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
asgiref=3.5.2=py310hca03da5_0
blosc=1.21.3=h313beb8_0
boost-cpp=1.78.0=hf1d6563_2
bzip2=1.0.8=h80987f9_5
c-ares=1.19.1=h80987f9_0
ca-certificates=2024.3.11=hca03da5_0
cairo=1.16.0=h73a0509_1014
cfitsio=4.2.0=h2f961c4_0
crispy-bootstrap5=2024.2=pyhd8ed1ab_0
curl=8.2.1=h449679c_0
django=4.1=py310hca03da5_0
django-bootstrap5=24.1=pyhd8ed1ab_0
django-crispy-forms=2.1=pyhd8ed1ab_0
django-environ=0.11.2=pyhd8ed1ab_0
django-leaflet=0.29.1=pyhd8ed1ab_0
djangorestframework=3.15.1=pyhd8ed1ab_0
djangorestframework-gis=1.0=pypi_0
expat=2.6.2=h313beb8_0
font-ttf-dejavu-sans-mono=2.37=hd3eb1b0_0
font-ttf-inconsolata=2.001=hcb22688_0
font-ttf-source-code-pro=2.030=hd3eb1b0_0
font-ttf-ubuntu=0.83=h8b1ccd4_0
fontconfig=2.14.1=hee714a5_2
fonts-anaconda=1=h8fa9717_0
fonts-conda-ecosystem=1=hd3eb1b0_0
freetype=2.12.1=h1192e45_0
freexl=1.0.6=h1a28f6b_0
gdal=3.6.2=py310hc67b115_5
geos=3.11.1=hb7217d7_0
geotiff=1.7.1=hdcdc974_6
gettext=0.21.1=h0186832_0
giflib=5.2.1=h80987f9_3
hdf4=4.2.15=h1a38d6a_5
hdf5=1.12.2=nompi_h8968d4b_100
icu=70.1=h6b3803e_0
jpeg=9e=h80987f9_1
json-c=0.16=h1a28f6b_0
kealib=1.5.0=hfd766a6_0
krb5=1.20.1=h8380606_1
lcms2=2.15=h481adae_0
lerc=4.0.0=h9a09cb3_0
libblas=3.9.0=16_osxarm64_openblas
libcblas=3.9.0=16_osxarm64_openblas
libcurl=8.2.1=h0f1d93c_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libedit=3.1.20230828=h80987f9_0
libev=4.33=h1a28f6b_1
libffi=3.4.4=hca03da5_0
libgdal=3.6.2=hf58f65c_5
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libglib=2.76.1=h4646484_0
libiconv=1.17=h0d3ecfb_2
libkml=1.3.0=hc4d7c42_7
liblapack=3.9.0=16_osxarm64_openblas
libnetcdf=4.8.1=nompi_h2510be2_106
libnghttp2=1.52.0=h10c0552_1
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libpq=15.1=hbce9e56_3
librttopo=1.1.0=h844f84d_12
libspatialite=5.0.1=h14115fc_23
libsqlite=3.45.2=h091b4b1_0
libssh2=1.10.0=h449679c_2
libtiff=4.5.0=h5dffbdd_2
libwebp-base=1.3.2=h80987f9_0
libxml2=2.10.3=h67585b2_4
libzip=1.9.2=h96606af_1
libzlib=1.2.13=h53f4e23_5
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
ncurses=6.4=h313beb8_0
nspr=4.35=h313beb8_0
nss=3.89.1=h313beb8_0
numpy=1.24.2=py310h3d2048e_0
openjpeg=2.5.0=hbc2ba62_2
openssl=1.1.1w=h1a28f6b_0
pcre2=10.40=hb34f9b4_0
pillow=10.3.0=pypi_0
pip=24.0=pypi_0
pixman=0.40.0=h1a28f6b_0
poppler=23.01.0=h9564b9f_0
poppler-data=0.4.11=hca03da5_1
postgresql=9.6.6=h65cfe13_1
proj=9.1.1=h13f728c_2
psycopg2=2.9.3=py310hfc29c8b_2
python=3.10.0=h70c1b39_3_cpython
python_abi=3.10=4_cp310
readline=8.2=h1a28f6b_0
sqlite=3.41.2=h80987f9_0
sqlparse=0.4.4=py310hca03da5_0
tiledb=2.13.2=hc7ac4c9_0
tk=8.6.12=hb8d0fd4_0
tzdata=2024a=h04d1e81_0
uriparser=0.9.7=h80987f9_0
xerces-c=3.2.4=h627aa08_1
xz=5.4.6=h80987f9_0
zlib=1.2.13=h53f4e23_5
zstd=1.5.5=hd90d995_0
